Flash spdiff with esp idf

+++ main/rtsp/rtsp_server.c
@@ -170,9 +170,11 @@ static void client_task(void *pvParameters) {
   if (getpeername(slot->socket, (struct sockaddr *)&peer_addr, &peer_len) ==
       0) {
     conn->client_ip = peer_addr.sin_addr.s_addr;
-    ESP_LOGI(TAG, "Client IP: %d.%d.%d.%d", conn->client_ip & 0xFF,
-             (conn->client_ip >> 8) & 0xFF, (conn->client_ip >> 16) & 0xFF,
-             (conn->client_ip >> 24) & 0xFF);
+    ESP_LOGI(TAG, "Client IP: %u.%u.%u.%u",
+             (unsigned int)(conn->client_ip & 0xFF),
+             (unsigned int)((conn->client_ip >> 8) & 0xFF),
+             (unsigned int)((conn->client_ip >> 16) & 0xFF),
+             (unsigned int)((conn->client_ip >> 24) & 0xFF));
   }
 
   // Allocate buffer
